requirements.txt
asgiref==3.8.1
Django==5.2.1
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
